+++ src/Main.hs
@@ -1,0 +1,5 @@
+module Main where
+
+import SHA1 as SHA1
+
+main = putStrLn SHA1.main