+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 ## 0.0.0 -- Jan. 26, 2023
 
 * First version that works. Nothing else besides.
+
+## 0.0.1 -- Jan. 30, 2023
+
+* Added "main" for profiling.
+* Fixed typo in "makeNice". 
+* Added internal function "scanWithHistory".
+* Refined message schedule algorithm. `main` now takes 2 seconds instead of 80.